requirements.txt
# config libs
configparser
# http libs
aiohttp
urllib3
# db libs
pymongo
# social networks libs
tweepy
foursquare
# scientific libs
nltk
numpy
matplotlib
scipy
# others
geopy
python-docx
manage.py
bson